add refresh value option

+++ src/v2/delete.ts
@@ -1,5 +1,5 @@
 import { Observable, from, throwError, isObservable } from 'rxjs'
-import { catchError, map } from 'rxjs/operators'
+import { catchError, map, take, tap } from 'rxjs/operators'
 import merge from 'lodash.merge'
 import { ctrl } from './__ctrl'
 import { handleNext } from './__util/handle-next'
@@ -7,21 +7,25 @@ import cloneDeep from 'lodash.clonedeep'
 import { optionsKey } from './__key'
 import reject from 'lodash.reject'
 import get from 'lodash.get'
+import { methods } from './__util/methods'
+import { LoadOptions } from './load'
 
 export class DeleteOptions {
   loadNext?: boolean
   deleteAll?: boolean
+  refreshValue?: boolean
   constructor(value?: DeleteOptions) {
     merge(this, value)
   }
 }
 
-const defaultOptions = new DeleteOptions({
+const defaults = new DeleteOptions({
   loadNext: null,
-  deleteAll: null
+  deleteAll: null,
+  refreshValue: null
 })
 
-export function Delete({ loadNext = defaultOptions.loadNext, deleteAll = defaultOptions.deleteAll } = defaultOptions) {
+export function Delete({ loadNext = defaults.loadNext, deleteAll = defaults.deleteAll, refreshValue = defaults.refreshValue } = defaults) {
   return function (target: any, propertyKey: string | symbol, descriptor: PropertyDescriptor) {
     const original = descriptor.value
     descriptor.value = function () {
@@ -29,7 +33,14 @@ export function Delete({ loadNext = defaultOptions.loadNext, deleteAll = default
         instanceCtrl = ctrl(this),
         { value, deleting, deletingSuccess } = instanceCtrl,
         returned = original.apply(this, arguments),
-        count = 1
+        count = 1,
+        refresh = () => {
+          const method = methods(this).find(key => get(this, [key, optionsKey]) instanceof LoadOptions)
+          if (method !== null && method !== undefined) {
+            const returned = this[method]()
+            if (isObservable(returned)) returned.pipe(take(1)).subscribe()
+          }
+        }
       const
         deleteValue = result => {
           if (get(target.constructor[optionsKey], `type`, Object) === Array && result && !deleteAll) {
@@ -44,14 +55,13 @@ export function Delete({ loadNext = defaultOptions.loadNext, deleteAll = default
           }
         },
         dial = () => <T>(src: Observable<T>) => src.pipe(
-          map(result => {
+          tap(result => {
             if (++count < 2) {
-              deleteValue(cloneDeep(result))
+              if (!refreshValue) deleteValue(cloneDeep(result))
               deleting.next(false)
               deletingSuccess.next(true)
-              return cloneDeep(result)
+              if (refreshValue) refresh()
             }
-            else return result
           }),
           catchError(response => {
             if (++count < 2) {
@@ -74,8 +84,13 @@ export function Delete({ loadNext = defaultOptions.loadNext, deleteAll = default
         }
         returned = returned.pipe(dial())
       }
-      else if (returned !== undefined) deleteValue(returned)
-      else { }
+      else if (returned !== undefined) {
+        if (!refreshValue) deleteValue(returned)
+        else refresh()
+      }
+      else {
+        if (refreshValue) refresh()
+      }
       return returned
     }
     descriptor.value[optionsKey] = new DeleteOptions({ loadNext, deleteAll })
